hash mainfile template with files (#18)
This change adds the mainfile template to the list of things we
hash to create the compiled binary name. This ensures that we
don't have to remember to bump the mageVer when we change the
template, it'll just always auto-update. mageVer only needs
to change if there's processing that will produce different output
from the template without the template actually changing