Add mg.Target
This function allows specifying files/directories as targets and returns true when the targets are met and false otherwise.

Targets are met if the ModTime of src is older than any ModTime of any  of the specified targets.

For recursive checks of directory with its file contents set recusive argument to true