remove mg dep from mainfile
We now create the context in the mainfile, so as to remove any dependency on
the mg library, since the user may have an older version vendored, that doesn't
have the Context function.

ALso added a test to ensure that we don't import any non-stdlib libraries in the
mainfile to avoid problems like this in the future.