Fix for issue #122 in github.com/magefile/mage
target.Path() and target.Dir() return an os.ErrNotExist
when the dst does not exist. They also return the same
error when any of the sources don't exist. This prevents
triggering an action when EITHER the target doesn't exist
or is older than some of the sources.

test cases were updated as well.