Set GOARCH to amd64 for TestListMagefilesIgnoresRespectsGOOSArg
The TestListMagefilesIgnoresRespectsGOOSArg test tries to list go
files with a different GOOS arg by setting GOOS to windows on Linux
systems and vice versa.  This test works fine on amd64, but breaks on
arm64 because a windows/arm64 pair is not available yet and as a
result, the `go list` command underlying the `Magefiles` function
fails.

Fix this problem by keeping GOARCH as amd64 so that the go list
command does not fail.