Detect deps that have already been invoked
Detect when a dependency was accidentally already invoked

mg.Deps(MyDep())

Normally this causes a nil pointer panic when we invoke the function

Error: runtime error: invalid memory address or nil pointer dereference

Now it will print the following as soon as its detected

Error: A dependency of the current target was defined improperly, with parenthesis. Dependencies should be defined as mg.Deps(MyDep), not mg.Deps(MyDep())

I have applied this change to checkfn so that it applies to serial and
parallel deps.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>